refactor: SignIn 관련 클래스 리팩토링
- import 경로 정리
- Optional을 활용한 null 처리 로직 개선
